feat/statistics : 최근 강의 목록보기

+++ FE/utf/src/components/main/ProfileSetting.vue
@@ -3,11 +3,20 @@
   <div class="container">
     <div :class="{ profileBox: !darkMode, profileBoxDark: darkMode }">
       Edit profile<br />
-      name<br /><input v-model="userName" /><br />
-      email<br /><input v-model="email" /><br />
-      <button type="button" class="btn btn-danger" style="margin-right: 50px" @click="deleteUser()">탈퇴</button>
-      <button type="button" class="btn btn-primary" @click="updateUser()">수정</button>
-      <button type="button" id="cancleBtn" class="btn btn-secondary">취소</button>
+      <label for="profileInputName">name</label><br />
+      <input id="profileInputName" v-model="userName" /><br />
+      <label for="profileInputEmail">email</label><br />
+      <input id="profileInputEmail" v-model="email" /><br />
+      <div class="btnGroup mt-3">
+        <button type="button" class="blueBtn" @click="updateUser()">수정</button>
+        <button type="button" class="redBtn" @click="cancel()">취소</button>
+      </div>
+      <b-row>
+        <b-col sm="1"></b-col>
+        <b-col sm="2">
+          <p @click="deleteUser()">탈퇴</p>
+        </b-col>
+      </b-row>
     </div>
   </div>
 </template>
@@ -27,17 +36,17 @@ export default {
       email: store.state.userInfo["email"],
     };
   },
-  mounted() {
-    const cancleBtn = document.querySelector("#cancleBtn");
-    cancleBtn.onclick = () => {
+  methods: {
+    cancel() {
       this.userName = store.state.userInfo["userName"];
       this.email = store.state.userInfo["email"];
-      this.$emit("cancle");
-    };
-  },
-  methods: {
+      this.$emit("cancel");      
+    },
     deleteUser() {
-      deleteUserRequest(store.state.userInfo["userId"]);
+      let flag = confirm("정말로 탈퇴 하시겠습니까??");
+      if(flag){
+        deleteUserRequest(store.state.userInfo["userId"]);
+      }
     },
     async updateUser() {
       await updateUserRequest(store.state.userInfo["userId"], this.userName, this.email);
@@ -48,4 +57,24 @@ export default {
 };
 </script>
 
-<style></style>
+<style scoped>
+#profileInputName, #profileInputEmail {
+  width: 80%;
+  border: none;
+  border-radius: 10px;
+  margin: auto;
+  background-color: #e2ebff;
+}
+p {
+  cursor: pointer;
+  font-size: 12px;
+  color: red;
+}
+.btnGroup {
+  display: flex;
+  justify-content: flex-end;
+}
+.btnGroup button {
+  margin-right: 10px;
+}
+</style>
